backend/requirements.txt
fastapi[standard]==0.112.2
uvicorn==0.30.6
requests==2.32.3
python-dotenv==1.0.1
osmnx==1.9.1
google-genai>=1.0.0
Pillow>=10.0.0
